Refactor glance parser reference
