+++ lib/glance.js
@@ -26,7 +26,7 @@ var _setStrategies = require('./set-strategies');
 
 var _setStrategies2 = _interopRequireDefault(_setStrategies);
 
-var _glanceParser = require('./glance-parser');
+var _glanceParser = require('../lib/glance-parser');
 
 var _glanceParser2 = _interopRequireDefault(_glanceParser);
