+++ package.json
@@ -4,7 +4,7 @@
   "description": "Automation at a glance",
   "main": "lib/glance.js",
   "scripts": {
-    "build": "babel src -d lib; node_modules/.bin/pegjs src/glance.pegjs lib/glance-parser.js",
+    "build": "node_modules/.bin/pegjs src/glance.pegjs lib/glance-parser.js; babel src -d lib;",
     "test": "npm run build; node_modules/.bin/wdio wdio.conf.js"
   },
   "repository": {
